First cut at j7 integration into MallocPlus.
This is really rough, but it doesn't crash. Please have a close look at all the
alloc functions I touched in this commit; I have have overdone the LOAD_BALANCE
stuff trying to get this to run. At least we know the base infrastructure is
okay. Lots of optimizations are still required. Furthermore, we may have some
allocs in there that are load balance even if the capability isn't present. This
is okay, because the allocs will just fall through, but I still don't like it.
At least I should be consistent; some LOAD_BALANCE alloc are
protected with "if defined HAVE_J7," while so aren't. Sorry for the mess.