Don't build with experimental software by default.
Only use quo and j7 if all requirements (Boost, MPI, OpenMP, QUO) are met. One
can easily toggle between clamr_quo builds and others by unsetting $QUO_HOME. We
want a "clean" base so we can make sense of what's going on with the experimental
infrastructure. So, l7 will NOT be built with QUO support, even if the minimum
QUO requirements are met; it will only be built with QUO support if all
requirements are met.