The prototype iterative hashing scheme is being added to the main branch from the iterative branch as a temporary backup. All edits have currently been put in state.cpp, and thus the added file is experimental_iterative_ state.cpp.
Currently, all code is resident in State::calc_finite_difference(), however
the vast majority will eventually be moved to the Mesh class.

This code is still experimental. In particular, the type computable_t will
eventually be completely hidden from the state kernel. And this reflects the
current experimental state of the code: HALF_COMPUTE cells are computed as if
they were COMPUTE cells. This will most likely lead to segmentation faults and
inconsistencies. When this is patched, the code will be ready for a merge into
the real state.cpp with the removal of rMesh into the Mesh class to follow.